backend/pyproject.toml
[project]
name = "agentboard"
version = "0.1.0"
description = "AgentBoard - Task Management System"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.110.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.27.0",
    "aiohttp>=3.9.0",
    "python-slugify>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
